feature(api): cart manipulations are performed through the auth token
