+++ api/cart-service/internal/router/router.go
@@ -10,9 +10,9 @@ func CreateRouter(h *handler.CartHandler) *mux.Router {
 	r := mux.NewRouter()
 
 	r.HandleFunc("/cart", h.GetCart).Methods("GET")
-	r.HandleFunc("/cart/{user_id}/products", h.AddProductToCart).Methods("POST")
-	r.HandleFunc("/cart/{user_id}/products/{product_id}", h.UpdateProductQuantity).Methods("PUT")
-	r.HandleFunc("/cart/{user_id}/products/{product_id}", h.RemoveProductFromCart).Methods("DELETE")
+	r.HandleFunc("/cart/products", h.AddProductToCart).Methods("POST")
+	r.HandleFunc("/cart/products/{product_id}", h.UpdateProductQuantity).Methods("PUT")
+	r.HandleFunc("/cart/products/{product_id}", h.RemoveProductFromCart).Methods("DELETE")
 
 	return r
 }
